Moves Table features to row three of the TinyMCE editor
Putting the Table buttons on the second row of the kitchen sink made it very
wide, so that on certain themes, enabling the kitchen sink made the editor
window too wide for the containing element. Moving these buttons to row 3
is a way around this.

It also required some additional javascript to ensure that the Kitchen Sink
button toggled rows other than 2.