document unlocking bug fix
Document unlocking was a little buggy. The Docs would be unlocked when
the user clicked any 'a' tag (eg, even the TinyMC's 'HTML' and 'Visual'
links), even if they didn't lead to a different page.
Also, in Safari, the setup the function 'bp_docs_load_idle', which sets
up document unloading stuff, wouldn't get called until the user clicked
on the Tiny MCE editor (and thus document unlocking wouldn't happen
even when a user clicked on an 'a' tag.)
My solution was to instead trigger unlocking only when the page is
unloaded. This way it also catches when the user simply closes the
browser, or is directed away from the page by a javascript action, or
whatever. So far I tested it in Safari 5, Chrome Version 20.0.1132.57,
Firefox 7, and IE 9.