First pass at taxonomy implementation of Read permissions
Storing the Read setting for individual docs in the settings array only made
it next to impossible to filter queries based on the current user's
permissions. By mirroring that setting in a taxonomy, we can perform complex
tax queries for different kinds of users.

Still to do: write the migrator