Filter for access protection at pre_get_posts
Previously, access protection was handled in the BP_Docs_Query process. This
had the effect of protecting access properly when a query initiated with Docs,
but not otherwise (as in the case of search, etc).

This changeset moves access protection to pre_get_posts, so that it will be
fired off with every single query that could possibly return a Doc.