In bp_docs_set_associated_group_id(), wipe out existing terms when setting group
Originally, this function was architected to allow for Docs to be associated
with more than one group. However, this feature was never built into the UI
(and probably won't be anytime soon). This changeset means that it's only
possible to associate a Doc with a single group, at least as far as the
function bp_docs_set_associated_group_id() is concerned (unless you pass in
an array for the $group_id).

This prevents problems when the associated group has been changed (previously,
the second group was just being added, rather than replacing the original
group).

See #286