Refactors doc save process to ensure group associations can be unset
Previously, associated group data was saved only when a group ID was found in
the $_POST global. This meant that when no value was sent - you were attempting
to unset the group association - the value was not processed.

This changeset modifies the way associated group data is detected, so that
when no value is passed, a value of 0 is assumed; and when a value of 0 is
passed to bp_docs_set_associated_group_id(), all associations are deleted
from the database for that doc.

Fixes #286