Fix the wonky tabindex on Text mode when editing a Doc
There's some weirdness in the way that tabindexes are set up when using Text
mode (vs Visual mode) of the WP editor. Namely, when you hit Tab from the
previous field, you have to cycle through all the styling buttons before
getting to the actual doc content.

As a workaround, I've added some code that detects tab-outs from the
Permalink field, and when one is detected, it sends focus to the end of the
doc_content textarea.

Fixes #311. Props @dcavins for the original patch.