Fix activity action when posting a group doc.
The current codebase checks $query->item_type.  This variable is currently
empty as of v1.4.5.

This commit worksaround this by checking if the posted 'associated_group_id' field exists.  If it does, this is a good indicator
that a group is attached to this doc.

There is probably a better way to do this in the BP Doc query builder, but
I'm not too familiar with the BP Doc codebase yet.