Fix "minimum role to associate with group" setting
Previously, when Doc creation happened within groups, it was possible
for the group admin to limit who was able to create Docs within the
group, by group role. Once Docs were moved out of groups, and "group Doc
creation" became "association of a Doc with a group", this functionatily
broke. The current changeset reintroduces it.

Fixes #233