Overhaul user- and group-specific tag cloud treatment
Previously, these items had been cached in groupmeta/usermeta, and then
reassembled when the page was loaded. This process was complex and full of
bugs, to the extent that it didn't work with users at all.

The new technique involves using WP_Query to pull up associated documents
inline, rather than attempting to cache them. This technique is more reliable,
and shouldn't cause overhead, since the same queries need to be performed later
on the directory page anyway (and are thus non-persistently cached by WP).

See #379

Props @dcavins for an initial patch