Improved compatibility with BP 2.0 groups_get_groupmeta() functions
In the previous behavior, groups_*_groupmeta() was stripping the hyphen from
the 'bp-docs' meta_key. On the upgrade to BP 2.0, this bug was fixed. But now,
calls to groups_get_groupmeta() were not finding the proper value (which had
been stored as 'bpdocs'). The new wrapper function ensures that legacy values
will still be found, while they get migrated over to the new settings format.