Better parsing of submitted bp-docs settings in groups
bp-docs settings include group-enable, which is represented in the markup by
a checkbox. This means that when the checkbox is unchecked, no value for group-
enable shows up in the POST array. As a result, when Docs is disabled for a
group, no value is saved in the settings array. Since the introduction of
bp_docs_get_group_settings() (and its internal default parsing), this has meant
that groups without a value for group-enable incorrectly get Docs turned on.

We fix this by ensuring that all bp_docs_get_group_settings() keys are filled
in saved settings arrays.

Fixes #424