Allow group admins/mods to remove doc from group.
Only site admins (users with ‘bp_moderate’ cap) and the doc owner can
change group association. This patch proposes a checkbox for removing a
specific group association from a doc. There is still a problem of a
redirect issue after save—the editing user no longer has access to the
doc, but is redirected to it, resulting in a bp_noaccess.