Assume BP Docs is not enabled unless explicitly enabled.
When BuddyPress Docs, via `bp_docs_get_group_settings()`, cannot find
the “bp-docs” groupmeta entry for a group, it assumes that the plugin
is active for the group. This causes poor behavior when the groupmeta
doesn’t exist. An example is when the plugin is activated after groups
are created.

Also remove some extra settings checks that are not needed because
default values are set in `bp_docs_get_group_settings()`.