Use bp_docs_locate_template() to select folder AJAX template.
When an AJAX request is made for the contents of a group folder,
select the template using `bp_docs_locate_template()`. Then, the
included template is easily overridden by creating a new file at
`/docs/docs-folder-loop.php` in the theme. Also,
`bp_docs_locate_template()` allows plugins and developers to filter the
location of the template for other situations.

Fixes #511.