Avoid duplicate div#buddypress when theme compatibility is active.
Many of the BP Docs templates were born in the dark times before
BuddyPress theme compatibility, so they include a wrapper div with the
id `buddypress`.  When used with theme compatibility, though, this
results in duplicate divs sharing the same ID. This change set adds a
check for the theme compatibility status before adding the extra
elements.