Use post__not_in for comment protection.
Passing `array()` or `array( 0 )` as the `post__in` argument just means
that the `post__in` query clause isn't calculated in
`WP_Comment_Query::get_comment_ids()`. The same is true when passing a
`post_id` of 0. So, pass all of the restricted comment ids to the
`post__not_in` argument instead.

Add tests that check the behavior of `get_comment()`, too. (Thanks
@r-a-y for the example comment tests to reference.)