add filter for group extension enable_edit_item
This allows plugins to override the default of showing
the docs settings tab in the group manage screen.

This is mainly for the purpose of forcing bp docs enabled
or disabled for some groups.

There is alredy a filter to hide the bp docs create step.
Adding a filter also for the edit screen is only consistent.

This does not change the default behaviour as it is now.