Prepare for WordPress.org translation support.
This commit attempts to prepare the plugin for compliance with
translate.wordpress.org requirements:
• Declare the text domain ‘buddypress-docs’ (must match the plugin’s
slug on WP.org) in the plugin header.
• Use the new text domain in all translation strings.
• Rename all language files to use the new text domain.
• Understand the enforced load order of language files as of WP 4.6,
which is that language files in
WP_LANG_DIR/plugins/buddypress-docs-LOCALE.mo are loaded automatically
(and before `init`), then our call to `load_plugin_textdomain()` would
find language files in
`buddypress-docs/languages/buddypress-docs-LOCALE.mo` if no match was
found in the global languages directory.