Folder management: Show "delete" link only when user can delete.
The “Delete” link is being included in the folder editing screen for
all users, but not every user who has the `bp_docs_manage_folders`
capability generally has the `bp_docs_manage_folder` capability for a
specific folder, which is required in
`bp_docs_process_folder_delete_cb()`. The link is now only generated
when the user actually could delete the folder.

See
https://wordpress.org/support/topic/user-group-cant-delete-doc-folder/