Fix attachment cache behavior.
The function `bp_docs_has_docs()` will preload the fetched docs' attachment caches when the argument `update_attachment_cache` is set to true. This is great and cuts down on queries. However, the call to `get_posts()` that fetches the attachments didn't specify that the query should fetch all matches, not just the first page of results. So, if the set of docs altogether include more than 20 attachments, some would be left off and the cached values would not be complete.