Add Akismet filtering of doc content.
If Akismet is configured for the site, send the content over to Akismet for inspection. If Akismet responds that the content looks like spam, set the post status to a new status, "bp_docs_pending," so that site admins can choose to either publish or delete the new content. Docs that are pending are visible to the creator and site admins only.