Add drag-and-drop folder doc-folder association.
Add support for moving docs between folders in the table view using drag and drop. Docs may moved into a folder (or subfolder) or dropped out to the top level (in no folder). Adapts code that was already in folders.js for the "tree" view.

Todos: Add some sort of visual indicator that a doc is draggable.