Improve access-related redirects for logged-in users.
Since https://buddypress.trac.wordpress.org/changeset/11610, logged-in
visitors to wp-login.php are redirected by BuddyPress to the URL indicated
by the `redirect_to` parameter. This change conflicts with Docs's behavior,
which is to send non-authorized (but authenticated) users to wp-login.php.
The current change works around this problem by sending logged-in users
to the home page instead.