History: Fix improper redirection for group docs for BuddyPress 12.
The history addon does a check for the doc's post revision before
allowing viewing. During the post revision check, the revision's post
parent can be zero.

In BuddyPress 12, when viewing a group doc, if the revision's post
parent is zero, combined with the get_post( 0 ) check, this can
actually return the post for the group directory instead of a falsey
value. This results in a redirect with 'edit.php?post_type=buddypress'
appended to the group doc's URL, which throws a 404.