Update the logic used to determine when Docs needs to modify post queries.
For privacy reasons, buddypress-docs intervenes in certain `WP_Query` instances,
to ensure that Docs are not returned to which the user doesn't have access.
In certain cases, it's possible to know for certainty that a query will match
`bp_doc` objects, and thus should be modified. In certain other cases, it's
possible to know for certainty that a query will *not* match any such objects,
and so does not need the (potentially expensive) modification. In between,
we previously took a fairly conservative approach, by appending a `post__not_in`
clause to the query when we were not certain either way - specifically when
no `post_type` was specified.

This approach caused a number of problems. First, because the access query
modification was broadly applied, there were performance considerations,
particularly when the query was clearly not needed - for example, during the
primary query on the home page. Second, the specific method for building
the `post__not_in` clause - which internally uses a `WP_Query` - caused
the main `WP_Query` object's global state to be modified in certain ways
that were difficult to predict. Specifically, the `is_home` flag sometimes
was set to true in these cases, causing unwanted behavior on group pages
and elsewhere in BuddyPress.

We change the approach so that we only modify queries when we are fairly
certain that they will, or could, return `bp_doc` objects. This includes:
- Cases where `post_type` is, or includes, `bp_doc`
- Cases where `post_type` is explicitly set to `any`
- Cases where a `bp_docs_tag` tax query is present

See #757.